Saving the game works now
Use GameControl functions Save() and Load() to save and load the game.

Right now the only variables are exp, level, and name.

GameControl has "DontDestroyOnLoad" so a single instance of it will exist throughout the game.

I also got this plugin called "TextMesh Pro" because it was free and added a bunch of features to text. Could be cool.